Implement Set<T> using vector
Not super happy with this, but as a first foray with the
language this was productive and successful. I managed to
get into a lot of problems writing this and successfully
found a way to get the rust compiler to not be SUPER unhappy
with what I wrote.

However I have my doubts about the rampage use of references.

Also I wish it were possible (dit easier) to do this with a
fixed-length array. Vecs make a lot of what we need to do
a lot easier and that's not the goal of this exercise.